modify del_dev_suite script so images are only deleted when "force" parameter is used, image deletion normally should not be necessary and takes long to pull them again from the web

+++ del_dev_suite.sh
@@ -7,7 +7,11 @@ docker-compose rm web
 docker-compose rm db
 
 docker ps -a | awk '{print $1}' | xargs --no-run-if-empty docker rm -f
-docker rmi $(docker images -q)
+
+if [ "$1" == "force" ]; then
+   echo "== deleting also images =="
+   docker rmi $(docker images -q)
+fi
 
 sudo rm -rf */migrations
